+++ lib/garner/strategies/context/key/caller.rb
@@ -36,7 +36,7 @@ module Garner
           # @param identity [Garner::Cache::Identity] The cache identity.
           # @param ruby_context [Object] An optional Ruby context.
           # @return [Garner::Cache::Identity] The modified identity.
-          def self.apply(identity, ruby_context = self)
+          def self.apply(identity, ruby_context = nil)
             value = nil
 
             if ruby_context.send(:caller)